requirements.txt
pyrogram
tgcrypto
moviepy
pillow
python-dotenv
psycopg2-binary
